Refactor CRUD User

+++ JobIT/src/main/java/fit/hcmute/JobIT/service/impl/UserServiceImpl.java
@@ -1,17 +1,20 @@
 package fit.hcmute.JobIT.service.impl;
 
-import fit.hcmute.JobIT.dto.response.Meta;
-import fit.hcmute.JobIT.dto.response.ResultPaginationResponse;
+import fit.hcmute.JobIT.converter.UserMapper;
+import fit.hcmute.JobIT.dto.response.*;
 import fit.hcmute.JobIT.entity.User;
+import fit.hcmute.JobIT.exception.IdInvalidException;
 import fit.hcmute.JobIT.repository.UserRepository;
 import fit.hcmute.JobIT.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 
 @Service
@@ -19,15 +22,27 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
+    private final PasswordEncoder passwordEncoder;
 
     @Override
-    public User createUser(User user) {
-        return userRepository.save(user);
+    public CreateUserResponse createUser(User user) {
+        if (userRepository.existsByEmail(user.getEmail())) {
+            throw new IdInvalidException("Email already exists: " + user.getEmail() + ". Please use a different email.");
+        }
+
+        String hashedPassword = passwordEncoder.encode(user.getPassword());
+        user.setPassword(hashedPassword);
+
+        User savedUser = userRepository.save(user);
+        return userMapper.toCreateUserResponse(savedUser);
     }
 
     @Override
-    public User getUserById(Long id) {
-        return userRepository.findById(id).orElse(null);
+    public UserResponse getUserById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new IdInvalidException("User not found with ID: " + id));
+        return userMapper.toUserResponse(user);
     }
 
     @Override
@@ -41,24 +56,46 @@ public class UserServiceImpl implements UserService {
         meta.setPages(pageUser.getTotalPages());
         meta.setTotal(pageUser.getTotalElements());
 
+        List<UserResponse> userResponses = pageUser.getContent().stream().map(userMapper::toUserResponse).toList();
+
         resultPaginationResponse.setMeta(meta);
-        resultPaginationResponse.setResult(pageUser.getContent());
+        resultPaginationResponse.setResult(userResponses);
 
         return resultPaginationResponse;
     }
 
     @Override
-    public User updateUser(User user) {
-        return userRepository.save(user);
+    public UpdateUserResponse updateUser(User user) {
+
+        User userCurrent = userRepository.findById(user.getId())
+                .orElseThrow(() -> new IdInvalidException("User not found with ID: " + user.getId()));
+
+        userCurrent.setName(user.getName());
+        userCurrent.setGender(user.getGender());
+        userCurrent.setAddress(user.getAddress());
+        userCurrent.setAge(user.getAge());
+
+        User updatedUser = userRepository.save(userCurrent);
+
+        return userMapper.toUpdateUserResponse(updatedUser);
+
     }
 
     @Override
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new IdInvalidException("User not found with ID: " + id);
+        }
         userRepository.deleteById(id);
     }
 
     @Override
     public User getUserByEmail(String email) {
         return userRepository.findByEmail(email);
+    }
+
+    @Override
+    public boolean existsByEmail(String email) {
+        return userRepository.existsByEmail(email);
     }
 }
